+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.8.0 (WIP)
 
 * Add ability to programmatically inject code into the client-side script
+* Upgrade `json-dry` dependency to v2.0.0
 
 ## 0.7.27 (2022-12-23)