+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Add ability to programmatically inject code into the client-side script
 * Upgrade `json-dry` dependency to v2.0.0
 * Add `Blast.createStaticDefiner` and `Blast.createProtoDefiner`
+* Remove `WeakMap` polyfill
 
 ## 0.7.27 (2022-12-23)