+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Remove old `String` polyfills
 * Remove `Symbol` polyfill
 * Clean up string-entities code
+* Remove `Blast.uneval` function
 
 ## 0.7.27 (2022-12-23)