+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Remove `Function#listenTo` and `Function#unlisten`
 * Remove `Object.values()` and `Object.entries()` polyfills
 * Remove `Object.zip()` and `Object.unzip()`
+* Remove `Object.assign()` polyfill
 
 ## 0.7.27 (2022-12-23)